requirements.txt
anyio==4.2.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
listify==2020.12.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
ollama==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.2
requests==2.31.0
runcmd==2020.12.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
values==2020.12.3
wcwidth==0.2.13
zipp==3.17.0
